move toplevel loki client
