+++ internal/component/loki/rules/lokiclient/client_test.go
@@ -1,4 +1,4 @@
-package client
+package lokiclient
 
 import (
 	"net/http"
