+++ internal/component/loki/rules/lokiclient/rules.go
@@ -1,4 +1,4 @@
-package client
+package lokiclient
 
 import (
 	"context"
