+++ internal/component/loki/rules/lokiclient/rules_test.go
@@ -1,4 +1,4 @@
-package client
+package lokiclient
 
 import (
 	"fmt"
